requirements.txt
streamlit-0.82.0-py2.py3-none-any.whl